[DTensor] Clean up ShardingProp LRUCache
This is not a very good PR.  Maybe we shouldn't land it as-is.  The
version that might be better is to expose a c++ entrypoint to
'propagate' that uses the c++ cache.  (Currently, the c++ cache is only
accessible via the dispatchDTensorOp codepath).

At least in this PR it is clear to readers that there is no more python
DTensor ShardingProp cache, and it is also more clear to the tail users
(sdpa, tp_conv) that they are using a separate cache that won't be
populated by / shared with propagation that happens during normal
DTensor dispatch.

Does this even matter?  Maybe it is fine for sdpa/tp_conv to have their
own separate python cache, if we do not expect to get reuse between
dispatch prop and sdpa/tp_conv prop.

[ghstack-poisoned]
